+++ src/pages/SignupPage.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import Signup from '../../components/landing/Signup';
-
+import Signup from '../components/landing/Signup';
 const SignupPage = () => {
   return <Signup />;
 };
