+++ src/router/Router.tsx
@@ -4,14 +4,13 @@ import { Routes, Route } from 'react-router-dom';
 import Aside from '../components/aside';
 import Layout from '../components/Layout';
 import Dashboard from '../pages/Dashboard';
-import LandingPage from '../pages/Landing/LandingPage';
+import LandingPage from '../pages/LandingPage';
 import AddAsset from '../pages/AddAsset';
 import ModifyAsset from '../pages/ModifyAsset';
-import ConfirmPage from '../pages/Signup/ConfirmPage';
-import SignupPage from '../pages/Signup/SignupPage';
-import EnterInfoPage from './../pages/Signup/EnterInfoPage';
-import FixPage from '../pages/Fix/FixPage';
-
+import ConfirmPage from '../pages/ConfirmPage';
+import FixPage from '../pages/FixPage';
+import SignupPage from '../pages/SignupPage';
+import EnterInfoPage from '../pages/EnterInfoPage';
 function Router() {
   return (
     <div>
